QtCreaterでのビルドエラーの修正。
git-svn-id: http://nasserver:888/repos/main/platform/hsp/tools/hspide@2372 fd21481f-4f5f-b64a-a7cc-914020ce4b36

+++ hspide/workspaceitem.cpp
@@ -19,7 +19,7 @@ CWorkSpaceItem::CWorkSpaceItem(QObject* parent, CWorkSpaceModel* model)
 	, m_buildConfigurations(theConf->buildConf())
 {
 	m_icon = QIcon(":/images/tango/small/text-x-generic.png");
-	m_text = QString("this=%1").arg((int)this, 0, 16);
+    m_text = QString("this=%1").arg((long long)this, 0, 16);
 
 	// V‚µ‚¢UUID‚ğŠ„‚è“–‚Ä
 	for(Configuration::BuildConfList::iterator
